Factored out PSF geometry calculation
