+++ specula/processing_objects/psf.py
@@ -1,5 +1,7 @@
 
 from specula import fuse
+from specula.lib.calc_psf_geometry import calc_psf_geometry
+
 from specula.base_processing_obj import BaseProcessingObj
 from specula.base_value import BaseValue
 from specula.data_objects.electric_field import ElectricField
@@ -27,31 +29,17 @@ class PSF(BaseProcessingObj):
                 ):
         super().__init__(target_device_idx=target_device_idx, precision=precision)
 
-        self.simul_params = simul_params
-        self.pixel_pupil = self.simul_params.pixel_pupil
-        self.pixel_pitch = self.simul_params.pixel_pitch
-        self.dim_pup_in_m = self.pixel_pupil * self.pixel_pitch
-
         if wavelengthInNm <= 0:
             raise ValueError('PSF wavelength must be >0')
         self.wavelengthInNm = wavelengthInNm
 
-        if nd is not None:
-            if pixel_size_mas is not None:
-                raise ValueError('Cannot set both nd and pixel_size_mas. Use one or the other.')
-            self.nd = nd
-        elif pixel_size_mas is not None:
-            self.nd = PSF.calc_psf_sampling(
-                self.pixel_pupil, 
-                self.pixel_pitch, 
-                self.wavelengthInNm, 
-                pixel_size_mas
-            )
-        else:
-            # Default case, use nd as a scaling factor
-            self.nd = 1.0
-        self.psf_pixel_size = self.calc_psf_pixel_size()
-
+        self.psf_pixel_size, self.nd = calc_psf_geometry(
+                                            simul_params.pixel_pupil,
+                                            simul_params.pixel_pitch,
+                                            wavelengthInNm,
+                                            nd,
+                                            pixel_size_mas)
+            
         self.start_time = start_time
 
         self.sr = BaseValue(target_device_idx=self.target_device_idx)
@@ -77,75 +65,6 @@ class PSF(BaseProcessingObj):
 
         self.out_size = [int(np.around(dim * self.nd/2)*2) for dim in in_ef.size]
         self.ref = Intensity(self.out_size[0], self.out_size[1], target_device_idx=self.target_device_idx)
-
-    def calc_psf_sampling(pixel_pupil: int, pixel_pitch: float, wavelength_nm: float, psf_pixel_size_mas: float):
-        """
-        Calculate PSF sampling parameters ensuring constraints are met
-
-        Args:
-            pixel_pupil: Number of pixels across the pupil
-            pixel_pitch: Physical size of each pixel in meters
-            wavelength_nm: Wavelength in nanometers
-            psf_pixel_size_mas: Desired PSF pixel size in milliarcseconds
-
-        Returns:
-            psf_sampling: The calculated sampling factor
-        """
-        
-        # Calculate pupil diameter in meters
-        dim_pup_in_m = pixel_pupil * pixel_pitch
-
-        # Calculate theoretical maximum pixel size (Nyquist limit)
-        max_pixel_size_mas = (wavelength_nm * 1e-9 / dim_pup_in_m * 3600 * 180 / np.pi) * 1000
-
-        if psf_pixel_size_mas > max_pixel_size_mas:
-            raise ValueError(
-                f"Requested PSF pixel size ({psf_pixel_size_mas:.2f} mas) is larger than "
-                f"the theoretical maximum ({max_pixel_size_mas:.2f} mas) for this wavelength and pupil size."
-            )
-
-        # Calculate required sampling
-        required_sampling = (wavelength_nm * 1e-9 / dim_pup_in_m * 3600 * 180 / np.pi) * 1000 / psf_pixel_size_mas
-
-        # Find nearest valid sampling (pixel_pupil * sampling must be integer)
-        # Try different integer values for the final PSF size
-        best_sampling = required_sampling
-        best_error = float('inf')
-
-        for psf_size in range(int(pixel_pupil * required_sampling) - 5, 
-                            int(pixel_pupil * required_sampling) + 6):
-            if psf_size > 0:
-                candidate_sampling = psf_size / pixel_pupil
-                candidate_pixel_size = max_pixel_size_mas / candidate_sampling
-                error = abs(candidate_pixel_size - psf_pixel_size_mas)
-
-                if error < best_error:
-                    best_error = error
-                    best_sampling = candidate_sampling
-
-        actual_psf_sampling = best_sampling
-        actual_pixel_size_mas = max_pixel_size_mas / actual_psf_sampling
-
-        # Warning if approximation is significant
-        error_percent = abs(actual_pixel_size_mas - psf_pixel_size_mas) / psf_pixel_size_mas * 100
-        if error_percent > 1.0:
-            print(f"Warning: Actual pixel size ({actual_pixel_size_mas:.2f} mas) differs from "
-                f"requested ({psf_pixel_size_mas:.2f} mas) by {error_percent:.1f}% due to "
-                f"integer sampling constraint.")
-
-        return actual_psf_sampling
-
-    def calc_psf_pixel_size(self):
-        """
-        Calculate PSF pixel size based on sampling factor or default settings.
-        
-        Returns:
-            pixel_size_mas 
-        """
-        
-        pixel_size_mas = (self.wavelengthInNm * 1e-9 / self.dim_pup_in_m * 3600 * 180 / np.pi) * 1000 / self.nd
-        
-        return pixel_size_mas
 
     def calc_psf(self, phase, amp, imwidth=None, normalize=False, nocenter=False):
         """
